Cleaned up and improved vispy integration.
Was having trouble with event handling and screen
updating using vispy. Learned a few new things and
integrated the framework a little better. It's a
bit messy, but that will be ok for now. Eventually
I would like to pare down the backend requirements
to just pyopengl and pyglfw, but my programmer
heart knows it's a waste of time for right now.